+++ lib/segment/src/vector_storage/turbo/read_only/read_ops.rs
@@ -1,34 +1,17 @@
-//! Read-only counterpart of [`TurboVectorStorage`](super::TurboVectorStorage).
-//!
-//! Opens the TurboQuant-encoded blob and deletion flags over an arbitrary
-//! [`UniversalRead`] backend (mmap / cache / remote), so a read-only segment can
-//! retrieve and score a `Turbo4`-typed vector storage. The quantizer is fully
-//! determined by `(dim, distance)` (fixed TQDT constants), so nothing beyond the
-//! encoded bytes and flags is read from disk.
-//!
-//! Scoring is reused verbatim from the writable storage via the shared
-//! [`TurboScoring`](super::TurboScoring) trait — this storage only provides the
-//! read surface it needs.
-
 use std::borrow::Cow;
-use std::path::Path;
 
 use common::bitvec::BitSlice;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::generic_consts::{AccessPattern, Random};
-use common::mmap::AdviceSetting;
-use common::sorted_slice::SortedSlice;
 use common::types::{PointOffsetType, ScoreType};
-use common::universal_io::{CachedReadFs, Populate, UniversalRead, UniversalReadFs, UserData};
+use common::universal_io::{UniversalRead, UserData};
 use quantization::EncodedStorage;
 use quantization::turboquant::EncodedQueryTQ;
-use quantization::turboquant::quantization::TurboQuantizer;
 use smallvec::{SmallVec, smallvec};
 
-use super::multi::{OFFSETS_PATH, TurboMultiScoring};
-use super::{DELETED_PATH, TQDT_BITS, TQDT_MODE, TQDT_ROTATION, TurboScoring, VECTORS_PATH};
-use crate::common::flags::in_memory_bitvec_flags::InMemoryBitvecFlags;
-use crate::common::live_reload::LiveReload;
+use super::super::TurboScoring;
+use super::super::multi::TurboMultiScoring;
+use super::{ReadOnlyTurboMultiVectorStorage, ReadOnlyTurboVectorStorage};
 use crate::common::operation_error::OperationResult;
 use crate::data_types::named_vectors::{CowMultiVector, CowVector};
 use crate::data_types::vectors::{
@@ -37,189 +20,11 @@ use crate::data_types::vectors::{
 use crate::spaces::metric::Metric;
 use crate::spaces::simple::{CosineMetric, DotProductMetric, EuclidMetric, ManhattanMetric};
 use crate::types::{Distance, MultiVectorConfig, VectorStorageDatatype};
-use crate::vector_storage::chunked_vectors::ChunkedVectorsRead;
 use crate::vector_storage::multi_dense::appendable_mmap_multi_dense_vector_storage::MultivectorMmapOffset;
-use crate::vector_storage::quantized::quantized_chunked_mmap_storage::QuantizedChunkedStorageRead;
-use crate::vector_storage::quantized::quantized_storage::QuantizedStorage;
 use crate::vector_storage::vector_storage_base::VectorStorageRead;
 use crate::vector_storage::{DenseTQVectorStorageRead, MultiTQVectorStorageRead, VectorOffsetType};
 
-/// Read-only TurboQuant encoded backend over a [`UniversalRead`] `S`.
-///
-/// Read-only counterpart of the writable
-/// [`TurboEncodedVectorStorage`](super::turbo_encoded_vectors::TurboEncodedVectorStorage):
-/// the storage layout is picked at open time from the segment's storage type.
-enum ReadOnlyTurboEncoded<S: UniversalRead> {
-    /// Single-file layout, written by the non-appendable `Mmap` / `InRamMmap`.
-    Single(QuantizedStorage<S>),
-    /// Chunked layout, written by the appendable `ChunkedMmap` / `InRamChunkedMmap`.
-    Chunked(QuantizedChunkedStorageRead<S>),
-}
-
-impl<S: UniversalRead> ReadOnlyTurboEncoded<S> {
-    fn get_quantized_vector(&self, key: PointOffsetType) -> Cow<'_, [u8]> {
-        match self {
-            Self::Single(s) => s.get_vector_data(key),
-            Self::Chunked(s) => s.get_vector_data(key),
-        }
-    }
-
-    fn get_quantized_vector_opt(&self, key: PointOffsetType) -> Option<Cow<'_, [u8]>> {
-        match self {
-            Self::Single(s) => s.get_vector_data_opt(key),
-            Self::Chunked(s) => s.get_vector_data_opt(key),
-        }
-    }
-
-    /// Run `f` for each vector in the batch, batching the underlying reads
-    /// where the backend supports it (the single-file backend pipelines
-    /// io_uring / prefetches mmap; the chunked backend batches through
-    /// [`EncodedStorage::for_each_batch`]).
-    fn for_each_in_batch<F: FnMut(usize, &[u8])>(
-        &self,
-        keys: &[PointOffsetType],
-        mut f: F,
-    ) -> OperationResult<()> {
-        match self {
-            Self::Single(s) => s.for_each_in_batch(keys, f),
-            Self::Chunked(s) => {
-                s.for_each_batch(keys, |idx, bytes| f(idx, &bytes));
-                Ok(())
-            }
-        }
-    }
-
-    fn vectors_count(&self) -> usize {
-        match self {
-            Self::Single(s) => s.vectors_count(),
-            Self::Chunked(s) => s.vectors_count(),
-        }
-    }
-
-    fn is_on_disk(&self) -> bool {
-        match self {
-            Self::Single(s) => s.is_on_disk(),
-            Self::Chunked(s) => s.is_on_disk(),
-        }
-    }
-
-    fn populate(&self) -> OperationResult<()> {
-        match self {
-            Self::Single(s) => {
-                s.populate();
-                Ok(())
-            }
-            Self::Chunked(s) => s.populate(),
-        }
-    }
-}
-
-/// Read-only TurboQuant dense vector storage over a [`UniversalRead`] backend.
-pub struct ReadOnlyTurboVectorStorage<S: UniversalRead> {
-    /// Read-only encoded vector blob over one of the two backends.
-    storage: ReadOnlyTurboEncoded<S>,
-    /// Quantizer used to de/quantize and score; rebuilt from `(dim, distance)`.
-    quantizer: TurboQuantizer,
-    /// Persisted soft-deletion flags, materialized in memory.
-    deleted: InMemoryBitvecFlags,
-    /// Distance used for scoring / query preprocessing.
-    distance: Distance,
-    /// Original (un-padded) vector dimensionality.
-    dim: usize,
-}
-
-impl<S: UniversalRead> std::fmt::Debug for ReadOnlyTurboVectorStorage<S> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ReadOnlyTurboVectorStorage")
-            .field("dim", &self.dim)
-            .field("distance", &self.distance)
-            .field("total_vector_count", &self.storage.vectors_count())
-            .field("deleted_count", &self.deleted.count())
-            .finish_non_exhaustive()
-    }
-}
-
 impl<S: UniversalRead> ReadOnlyTurboVectorStorage<S> {
-    /// Build the quantizer for a `Turbo4` storage; fully determined by
-    /// `(dim, distance)` and the fixed TQDT constants (matches the writable
-    /// `open_turbo_vector_storage_impl`).
-    fn build_quantizer(dim: usize, distance: Distance) -> TurboQuantizer {
-        TurboQuantizer::new(
-            dim,
-            TQDT_BITS,
-            TQDT_MODE,
-            distance.into(),
-            TQDT_ROTATION,
-            None,
-        )
-    }
-
-    /// Schedule background prefetch of the files [`Self::open`] will read.
-    ///
-    /// Absent files are skipped rather than reported: the subsequent open is
-    /// the one to produce the error.
-    pub fn preopen(
-        fs: &impl CachedReadFs<File = S>,
-        path: &Path,
-        chunked: bool,
-        populate: Populate,
-    ) -> OperationResult<()> {
-        let vectors_path = path.join(VECTORS_PATH);
-        if chunked {
-            QuantizedChunkedStorageRead::<S>::preopen(fs, &vectors_path, populate)?;
-        } else {
-            QuantizedStorage::<S>::preopen(fs, &vectors_path, populate)?;
-        }
-        InMemoryBitvecFlags::preopen(fs, &path.join(DELETED_PATH))?;
-        Ok(())
-    }
-
-    /// Open the read-only counterpart of a `Turbo4` dense storage at `path`,
-    /// threading every file open through `fs`; reads the existing layout but
-    /// creates and writes nothing. `chunked` selects the on-disk layout the
-    /// writable storage produced for the segment's storage type.
-    pub fn open(
-        fs: &impl UniversalReadFs<File = S>,
-        path: &Path,
-        dim: usize,
-        distance: Distance,
-        chunked: bool,
-        populate: Populate,
-    ) -> OperationResult<Self> {
-        let quantizer = Self::build_quantizer(dim, distance);
-        let quantized_vector_size = quantizer.quantized_size();
-        let vectors_path = path.join(VECTORS_PATH);
-
-        let storage = if chunked {
-            ReadOnlyTurboEncoded::Chunked(QuantizedChunkedStorageRead::open(
-                fs,
-                &vectors_path,
-                quantized_vector_size,
-            )?)
-        } else {
-            ReadOnlyTurboEncoded::Single(QuantizedStorage::from_file(
-                fs,
-                &vectors_path,
-                quantized_vector_size,
-            )?)
-        };
-
-        // The read-only backends map lazily; warm them when the load profile asks.
-        if !matches!(populate, Populate::No) {
-            storage.populate()?;
-        }
-
-        let deleted = InMemoryBitvecFlags::open::<S>(fs, &path.join(DELETED_PATH))?;
-
-        Ok(Self {
-            storage,
-            quantizer,
-            deleted,
-            distance,
-            dim,
-        })
-    }
-
     /// Whether scores must be negated to follow qdrant's "higher = better"
     /// convention (mirrors `TurboVectorStorage::invert_score`).
     fn invert_score(&self) -> bool {
@@ -375,120 +180,7 @@ impl<S: UniversalRead> TurboScoring for ReadOnlyTurboVectorStorage<S> {
     }
 }
 
-impl<S: UniversalRead> LiveReload for ReadOnlyTurboVectorStorage<S> {
-    type Fs = S::Fs;
-
-    /// Pick up vectors a writer appended (chunked backend only; the single-file
-    /// layout is immutable) and patch the in-memory deletion flags.
-    fn live_reload(
-        &mut self,
-        fs: &S::Fs,
-        deleted_points: &SortedSlice<'_, PointOffsetType>,
-        new_points: &SortedSlice<'_, PointOffsetType>,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()> {
-        if let ReadOnlyTurboEncoded::Chunked(storage) = &mut self.storage {
-            storage.live_reload(fs, deleted_points, new_points, hw_counter)?;
-        }
-        self.deleted.insert_all(deleted_points);
-        Ok(())
-    }
-}
-
-/// Read-only TurboQuant multivector vector storage over a [`UniversalRead`]
-/// backend. Read-only counterpart of
-/// [`TurboMultiVectorStorage`](super::multi::TurboMultiVectorStorage).
-///
-/// Multivectors are always stored in the appendable chunked layout, so — unlike
-/// the dense storage — there is a single backend.
-pub struct ReadOnlyTurboMultiVectorStorage<S: UniversalRead> {
-    /// Flat inner-vector space: one fixed-size encoded record per inner vector.
-    storage: QuantizedChunkedStorageRead<S>,
-    /// Maps each point to its record range in the inner space.
-    offsets: ChunkedVectorsRead<MultivectorMmapOffset, S>,
-    /// Quantizer used to de/quantize and score; rebuilt from `(dim, distance)`.
-    quantizer: TurboQuantizer,
-    /// Persisted soft-deletion flags, materialized in memory.
-    deleted: InMemoryBitvecFlags,
-    /// Distance used for scoring / query preprocessing.
-    distance: Distance,
-    /// Original (un-padded) inner vector dimensionality.
-    dim: usize,
-    multi_vector_config: MultiVectorConfig,
-}
-
-impl<S: UniversalRead> std::fmt::Debug for ReadOnlyTurboMultiVectorStorage<S> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ReadOnlyTurboMultiVectorStorage")
-            .field("dim", &self.dim)
-            .field("distance", &self.distance)
-            .field("total_vector_count", &self.offsets.len())
-            .field("deleted_count", &self.deleted.count())
-            .finish_non_exhaustive()
-    }
-}
-
 impl<S: UniversalRead> ReadOnlyTurboMultiVectorStorage<S> {
-    /// Schedule background prefetch of the files [`Self::open`] will read.
-    ///
-    /// Absent files are skipped rather than reported: the subsequent open is
-    /// the one to produce the error.
-    pub fn preopen(
-        fs: &impl CachedReadFs<File = S>,
-        path: &Path,
-        advice: AdviceSetting,
-        populate: Populate,
-    ) -> OperationResult<()> {
-        QuantizedChunkedStorageRead::<S>::preopen(fs, &path.join(VECTORS_PATH), populate)?;
-        ChunkedVectorsRead::<MultivectorMmapOffset, S>::preopen(
-            fs,
-            &path.join(OFFSETS_PATH),
-            advice,
-            populate,
-        )?;
-        InMemoryBitvecFlags::preopen(fs, &path.join(DELETED_PATH))?;
-        Ok(())
-    }
-
-    /// Open the read-only counterpart of a `Turbo4` multivector storage at
-    /// `path`, threading every file open through `fs`; reads the existing layout
-    /// but creates and writes nothing.
-    pub fn open(
-        fs: &impl UniversalReadFs<File = S>,
-        path: &Path,
-        dim: usize,
-        distance: Distance,
-        multi_vector_config: MultiVectorConfig,
-        advice: AdviceSetting,
-        populate: Populate,
-    ) -> OperationResult<Self> {
-        let quantizer = ReadOnlyTurboVectorStorage::<S>::build_quantizer(dim, distance);
-
-        let storage = QuantizedChunkedStorageRead::open(
-            fs,
-            &path.join(VECTORS_PATH),
-            quantizer.quantized_size(),
-        )?;
-        // The chunked read backend maps lazily; warm it when the profile asks.
-        if !matches!(populate, Populate::No) {
-            storage.populate()?;
-        }
-
-        let offsets = ChunkedVectorsRead::open(fs, &path.join(OFFSETS_PATH), 1, advice, populate)?;
-
-        let deleted = InMemoryBitvecFlags::open::<S>(fs, &path.join(DELETED_PATH))?;
-
-        Ok(Self {
-            storage,
-            offsets,
-            quantizer,
-            deleted,
-            distance,
-            dim,
-            multi_vector_config,
-        })
-    }
-
     /// Offset record for `key`, if the point exists.
     fn get_offset<P: AccessPattern>(&self, key: PointOffsetType) -> Option<MultivectorMmapOffset> {
         let record = self.offsets.get::<P>(key as VectorOffsetType)?;
@@ -765,26 +457,5 @@ impl<S: UniversalRead> TurboMultiScoring for ReadOnlyTurboMultiVectorStorage<S> 
             sum += max_sim;
         }
         sum
-    }
-}
-
-impl<S: UniversalRead> LiveReload for ReadOnlyTurboMultiVectorStorage<S> {
-    type Fs = S::Fs;
-
-    /// Pick up multivectors a writer appended (records + offsets) and patch the
-    /// in-memory deletion flags.
-    fn live_reload(
-        &mut self,
-        fs: &S::Fs,
-        deleted_points: &SortedSlice<'_, PointOffsetType>,
-        new_points: &SortedSlice<'_, PointOffsetType>,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()> {
-        self.storage
-            .live_reload(fs, deleted_points, new_points, hw_counter)?;
-        self.offsets
-            .live_reload(fs, deleted_points, new_points, hw_counter)?;
-        self.deleted.insert_all(deleted_points);
-        Ok(())
     }
 }
